moving all c/fortran cmake includes + lib linkings to own directory and cmake file
