+++ src/umpire/interface/CMakeLists.txt
@@ -4,6 +4,10 @@
 #
 # SPDX-License-Identifier: (MIT)
 ##############################################################################
+#if (ENABLE_FORTRAN OR ENABLE_C)
+#  add_subdirectory(c_fortran)
+#endif()
+
 add_shroud(
   YAML_INPUT_FILE umpire_shroud.yaml
   YAML_OUTPUT_DIR yaml
