+++ tests/integration/interface/allocator_c_tests.cpp
@@ -6,7 +6,7 @@
 //////////////////////////////////////////////////////////////////////////////
 #include "gtest/gtest.h"
 #include "umpire/config.hpp"
-#include "umpire/interface/umpire.h"
+#include "umpire/interface/c_fortran/umpire.h"
 
 static int unique_name = 0;
